chore: revert to old deploy script

+++ deploy.sh
@@ -1,12 +1,22 @@
 #!/bin/bash
 set -x
 
-if [ "$TRAVIS_JOB_NUMBER" = "$TRAVIS_BUILD_NUMBER.1" ]
+if [ "$TRAVIS_PULL_REQUEST" = "false" ] && [ "$TRAVIS_JOB_NUMBER" = "$TRAVIS_BUILD_NUMBER.1" ]
 then
-
-  echo "TEST: Deploying master to test server"
-  server="meditation-dev.sirimangalo.org"
-  version="$TRAVIS_BUILD_NUMBER.0.0"
+  # Determine folder and version based on tag or branch
+  if [ ! -z "$TRAVIS_TAG" ]
+  then
+    echo "PROD: Deploying $TRAVIS_TAG to server"
+    server="meditation.sirimangalo.org"
+    version=$TRAVIS_TAG
+  elif [ "$TRAVIS_BRANCH" = "master" ]
+  then
+    echo "TEST: Deploying master to test server"
+    server="meditation-dev.sirimangalo.org"
+    version="$TRAVIS_BUILD_NUMBER.0.0"
+  else
+    exit
+  fi
 
   chmod 600 deploy_key
   rm -f dist/src/config/config.json
